requirements.txt
Levenshtein
git+https://github.com/huggingface/transformers.git
datasets
lightning
peft
accelerate
bitsandbytes
wandb
python-dotenv
opencv-python==4.8.0.74
colorama
